(WIP) refactoring poll preview
