+++ web/src/App.tsx
@@ -2,7 +2,7 @@ import GlobalStyles from "./components/GlobalStyles";
 import { Header } from "./components/Header/Header";
 import { Footer } from "./components/Footer/Footer";
 import { Poll } from "./components/Poll/Poll";
-import { PollsPreviewContainer } from "./components/Poll/PollsPreviewContainer";
+import { PollsPreviewContainer } from "./components/Poll/PollPreview/PollsPreviewContainer";
 import styled from "styled-components";
 import { PollsContainer } from "./components/Poll/PollsContainer";
